% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltgwas.R
\name{ltgwas}
\alias{ltgwas}
\title{Association Teste using Linear Transformation of Genomic Evaluations}
\usage{
ltgwas(file = NULL, varG = NULL, varE = NULL, gebv = NULL)
}
\arguments{
\item{file}{name of all plink files (.bed, bim and .fam)}

\item{varG}{genetic variance}

\item{varE}{residual variance}

\item{gebv}{genomic estimated breeding value}
}
\value{
Returns SNP effects and SNP variance
}
\description{
Standardized test of association where genomic breeding values from a mixed model including fixed effects and a genomic information are linearly transformed to estimate marker effects and pvalues.
}
\examples{

# GWAS <- ltgwas(file = "filename,
#                varG=10,
#                varE=20,
#                gebv=gebv)

}
\author{
Fernando Brito Lopes
}
